#32363: Moved files and code to more logical places
