+++ src/github/interfaces.d.ts
@@ -120,3 +120,17 @@ export interface Comment {
   reactions?: Reactions;
   performed_via_github_app?: GitHubApp | null;
 }
+
+export interface ChangedFile {
+  sha: string;
+  filename: string;
+  status: 'added' | 'removed' | 'modified' | 'renamed' | 'copied' | 'changed' | 'unchanged';
+  additions: number;
+  deletions: number;
+  changes: number;
+  blob_url: string;
+  raw_url: string;
+  contents_url: string;
+  patch?: string | undefined;
+  previous_filename?: string | undefined;
+}
