+++ src/helper/interfaces.d.ts
@@ -72,20 +72,6 @@ export interface ReviewComments {
   unpostable: Annotation[];
 }
 
-export interface ChangedFile {
-  sha?: string;
-  filename: string;
-  status?: 'added' | 'removed' | 'modified' | 'renamed' | 'copied' | 'changed' | 'unchanged';
-  additions?: number;
-  deletions?: number;
-  changes?: number;
-  blob_url?: string;
-  raw_url?: string;
-  contents_url?: string;
-  patch?: string | undefined;
-  previous_filename?: string | undefined;
-}
-
 export interface AnalyzedFile {
   formattedValue: string;
   letter?: string;
