+++ test/github/objects/pulls.ts
@@ -1,0 +1,13 @@
+import { ChangedFile } from '../../../src/github/interfaces';
+
+export const changedFile: ChangedFile = {
+  sha: '',
+  filename: 'test.js',
+  status: 'renamed',
+  additions: 0,
+  deletions: 0,
+  changes: 1,
+  blob_url: '',
+  raw_url: '',
+  contents_url: ''
+};
