+++ test/github/review.test.ts
@@ -1,10 +1,10 @@
-import { githubConfig, octokit, ticsConfig } from '../../../src/configuration';
-import { postNothingAnalyzedReview, postReview } from '../../../src/github/posting/review';
-import { createSummaryBody } from '../../../src/helper/summary';
-import { Events } from '../../../src/helper/enums';
-import { logger } from '../../../src/helper/logger';
+import { githubConfig, octokit, ticsConfig } from '../../src/configuration';
+import { postNothingAnalyzedReview, postReview } from '../../src/github/review';
+import { createSummaryBody } from '../../src/helper/summary';
+import { Events } from '../../src/helper/enums';
+import { logger } from '../../src/helper/logger';
 
-jest.mock('../../../src/helper/summary', () => {
+jest.mock('../../src/helper/summary', () => {
   return {
     createSummaryBody: jest.fn()
   };
